Portability fixes for configure
The `test` command, as well as the `[` command, is not required to
know the `==` operator. Only a few implementations like bash and some
versions of ksh support it.

When you run `test foo == foo` on a platform that does not support the
`==` operator, the result will be false instead of true. This can lead
to unexpected behavior.